updated file
In this file there is a csv file created called Kaggle. We have to submit in the kaggle competition this field which is in csv format. It contains Gene ID and the probability that it is active. we are getting 88% accuracy ryt now but need to work on that . Kiitos :)